removing connect() option connectTimeout
We didn't realize that imap had a connection timeout in place
using the constructor option `authTimeout`, so we'll just use
that instead. `ConnectionTimeoutError` will continue to be the
error class the connect() method is rejected with when a
timeout occurs. The `connectTimeout` option is now deprecated
and  will still be supported until the next major release.