Add functionality to add and delete mailboxes
For these features the underlying 'node-imap' package
does not really support callbacks.
Thus callbacks are omitted for these.

Features are just promise wrappers to
the corresponding functionality of 'node-imap'.
Usage is quite simple and straighforward.
Either promise resolves to the requested boxname
or promise is rejected with error message from 'node-imap' .